Move `assertProtocolUrl` to urlUtils
