+++ src/utils/urlUtils.test.ts
@@ -15,9 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { assertProtocolUrl } from "@/errors/assertProtocolUrl";
 import { BusinessError } from "@/errors/businessErrors";
-import { makeURL, selectAbsoluteUrl } from "@/utils/urlUtils";
+import {
+  assertProtocolUrl,
+  makeURL,
+  selectAbsoluteUrl,
+} from "@/utils/urlUtils";
 
 describe("assertHttpsUrl", () => {
   test("parses HTTPS URLs", () => {
